Move parsing logic to separate procedure
The main benefit of separating the parsing from the interface script
is that now reading the input files and processing them can be done
in separate steps. By decoupling these two, it is easier to implement
a behaviour in which the parsing can make use of a Sound object if
available (to derive the timing of the final TextGrid object), but
can work well without it, which is ultimately more useful.

Additionally, this version of the parsing procedure has been greatly
expanded, to support multi-level and multi-alternative label files,
as explained in section 6.1 of the HTK Book. Multi-alternative label
files will generate a separate TextGrid for each. Multiple levels, on
the other hand, are stored in separate tiers.