Add MLF reader
This version of the reader follows the plan proposed in #11 [1]
and generates a MLF object representation using a Table. The
Table stores the pattern, mode, and target information, and
is used to provide results from queries made against it.

A querying script, which will contain the logic for the pattern
matching, is in development.

[1]: https://github.com/badner/lab_to_praat/issues/11